requirements.txt
flask
sqlalchemy
flask-sqlalchemy
flask-wtf
wtforms
gunicorn
flask-bootstrap4